requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.100.0
spotipy
python-dotenv
langchain